fix: Use new Bulk Downloads Endpoint (#294)
