+++ resolver-functions/fedBulkDownloads/fedBulkDownloads.test.ts
@@ -1,9 +1,9 @@
 import { ExecuteMeshFn } from "@graphql-mesh/runtime";
-import { getMeshInstance } from "./utils/MeshInstance";
+import { getMeshInstance } from "../../tests/utils/MeshInstance";
 
-import * as httpUtils from "../utils/httpUtils";
-import { getExampleQuery } from "./utils/ExampleQueryFiles";
-jest.mock("../utils/httpUtils");
+import * as httpUtils from "../../utils/httpUtils";
+import { getExampleQuery } from "../../tests/utils/ExampleQueryFiles";
+jest.mock("../../utils/httpUtils");
 
 beforeEach(async () => {
   (httpUtils.get as jest.Mock).mockClear();
@@ -70,41 +70,8 @@ describe("bulkDownloads Query:", () => {
 
   it("should give correct response with url params & successful run", async () => {
     const query = getExampleQuery("bulk-downloads-with-limit");
-    const railsResponse = [
-      {
+    const railsResponse = [{
         id: 12715,
-        params_json:
-          '{"sample_ids":{"value":[24682,25015,25160]},"workflow":{"value":"short-read-mngs"},"background":{"value":350,"displayName":"floo sp86"}}',
-        download_type: "sample_taxon_report",
-        status: "success",
-        error_message: null,
-        user_id: 412,
-        created_at: "2024-02-15T11:52:57.000-08:00",
-        updated_at: "2024-02-15T11:53:22.000-08:00",
-        progress: 0.87,
-        ecs_task_arn: null,
-        output_file_size: 57197,
-        description: null,
-        deleted_at: null,
-        analysis_type: "short-read-mngs",
-        analysis_count: 3,
-        num_samples: 3,
-        download_name: "Sample Taxon Reports",
-        file_size: "55.9 KB",
-        user_name: "Suzette McCanny",
-        execution_type: "resque",
-        log_url: null,
-        params: {
-          sample_ids: { value: [24682, 25015, 25160] },
-          workflow: ["short-read-mngs"],
-        },
-      },
-    ];
-    const railsResponse2 = {
-      bulk_download: {
-        id: 12715,
-        params_json:
-          '{"sample_ids":{"value":[24682,25015,25160]},"workflow":{"value":"short-read-mngs"},"background":{"value":350,"displayName":"floo sp86"}}',
         download_type: "sample_taxon_report",
         status: "success",
         error_message: null,
@@ -135,20 +102,8 @@ describe("bulkDownloads Query:", () => {
         ],
         workflow_runs: [],
         presigned_output_url: "https://presignedUrl.com",
-      },
-      download_type: {
-        type: "sample_taxon_report",
-        display_name: "Sample Taxon Reports",
-        description:
-          "Computed metrics (e.g. total reads, rPM) and metadata for each taxon identified in the sample",
-        category: "reports",
-        execution_type: "resque",
-        file_type_display: ".csv",
-        workflows: ["short-read-mngs"],
-      },
-    };
+    }];
     (httpUtils.get as jest.Mock).mockImplementationOnce(() => railsResponse);
-    (httpUtils.get as jest.Mock).mockImplementationOnce(() => railsResponse2);
     const result = await execute(query, {
       limit: 2,
       searchBy: "Suzette McCanny",
